# gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.0.0"
servers:
  - url: http://localhost:8083/portal/environments/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
...
  /apis/{apiId}:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/apiIncludeParam'
      summary: Get the API definition
      description:  |
        Get the detail of an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getApiByApiId
      security: []
      responses:
        200:
          description: API definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}/informations:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Portal
      summary: Get the api dynamic informations to display.
      operationId: getApiInformations
      description: |
        Get api informations.
      security: []
      responses:
        200:
          description: List of api informations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiInformation'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}/links:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      summary: Get the API links from Aside SYSTEM_FOLDER.
      operationId: getApiLinks
      description: |
        Get all the links (internal and external) to be displayed in the detail of the API.
      security: []
      responses:
        200:
          description: List of Links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinksResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /apis/{apiId}/pages:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/homePageQueryParam'
        - $ref: '#/components/parameters/parentPageQueryParam'
      summary: List API pages
      description: |
        List all documentation pages of an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getPagesByApiId
      security: []
      responses:
        200:
          $ref: '#/components/responses/PagesSuccess'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /apis/{apiId}/picture:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      summary: Get the API's picture
      description: |
        Get the API's picture.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getPictureByApiId
      security: []
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}/background:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      summary: Get the API's background
      description: |
        Get the API's background.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getBackgroundByApiId
      security: []
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}/plans:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'

      summary: List plans for an API
      description: |
        List plans for an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getApiPlansByApiId
      security: []
      responses:
        200:
          description: List accessible plans for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlansResponse'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}/ratings:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/mineQueryParam'
        - $ref: '#/components/parameters/ratingOrderQueryParam'
      summary: List ratings for an API
      description: |
        List ratings for an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getApiRatingsByApiId
      security: []
      responses:
        200:
          description: List of ratings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingsResponse'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Api
      summary: Create a rating for an API
      description: |
        Create a rating for an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.

        The current must have API_RATING[CREATE] permission to create a rating.
      operationId: createApiRating
      requestBody:
        description: Use to add a rating to an api
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingInput'
      responses:
        201:
          description: Created Rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        400:
          description: Rating for this api and this user already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Rating service is not activated or user does not have sufficient rights to access this resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /apis/{apiId}/subscribers:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/subscriptionStatusesQueryParam'

      summary: List applications that subscribred to an API
      description: |
        If the current user is the owner of the API, all connected applications will be returned. Filtered by status. Ordered by number of hits.
        Else only applications that current is allowed to access will.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getSubscriberApplicationsByApiId
      security: []
      responses:
        200:
          description: List accessible applications for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationsResponse'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /pages:
    get:
      tags:
        - Portal
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/homePageQueryParam'
        - $ref: '#/components/parameters/parentPageQueryParam'
      summary: List portal pages
      description: |
        List all portal documentation pages
      operationId: getPages
      security: []
      responses:
        200:
          $ref: '#/components/responses/PagesSuccess'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    apiIdParam:
      name: apiId
      in: path
      required: true
      description: Id of an API.
      schema:
        type: string
...
    mineQueryParam:
      name: mine
      in: query
      required: false
      description: If true, returns only connected user data.
      schema:
        type: boolean

    #Pagination
    pageNumberParam:
      name: page
      in: query
      required: false
      description: The page number for pagination.
      schema:
        type: integer
        default: 1
    pageSizeParam:
      name: size
      in: query
      required: false
      description: |
        The number of items per page for pagination.
        If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
        If the size is -1, the response contains all datas.
      schema:
        type: integer
        default: 10

    #Apis filtering
...
    apiIncludeParam:
      name: include
      in: query
      required: false
      description: Comma-separated list of related objects to include in the response.
      schema:
        type: array
        items:
          type: string
          enum:
            - pages
            - plans
    ratingOrderQueryParam:
      name: order
      in: query
      required: false
      description: |
        A  list of "field" elements.
        By default, sort is ASC. If *field* starts with '-', the order sort is DESC.\
        Currently, only **value**, **date** and **answers** are supported.
      example: "order=value"
      schema:
        type: string

    #Applications filtering
...
    homePageQueryParam:
      name: homepage
      in: query
      required: false
      description: If true, only the documentation homepage of the portal is returned.
      schema:
        type: boolean
    parentPageQueryParam:
      name: parent
      in: query
      required: false
      description: The name of the parent documentation page. If not null, only this page and its children are returned.
      schema:
        type: string
...
    subscriptionStatusesQueryParam:
      name: statuses
      in: query
      required: false
      description: status of subscription.
      schema:
        type: array
        items:
          type: string
          enum:
            - ACCEPTED
            - CLOSED
            - PAUSED
            - PENDING
            - REJECTED
...
  schemas:
...
    ErrorResponse:
      properties:
        errors:
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    ApplicationsResponse:
      properties:
        data:
          description: List of applications.
          type: array
          items:
            $ref: '#/components/schemas/Application'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    PagesResponse:
      properties:
        data:
          description: List of pages.
          type: array
          items:
            $ref: '#/components/schemas/Page'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
    RatingsResponse:
      properties:
        data:
          description: List of ratings.
          type: array
          items:
            $ref: '#/components/schemas/Rating'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
    PlansResponse:
      properties:
        data:
          description: List of data.
          type: array
          items:
            $ref: '#/components/schemas/Plan'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    LinksResponse:
      properties:
        slots:
          description: |
            Map of CategorizedLinks. Keys of the map can be:
            * aside
            * header
            * topfooter
            * footer
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/CategorizedLinks'
...
    Error:
      properties:
        status:
          description: Status HTTP of the error
          type: string
        message:
          description: Message of the error
          type: string
        code:
          description: Technical code of the error
          type: string
        parameters:
          description: Parameters
          type: object
          additionalProperties:
            type: string
    Api:
      description: |
        Describes an API.
      required:
        - id
        - name
        - version
        - description
        - owner
      properties:
        id:
          description: Unique identifier of an API.
          type: string
        name:
          description: Name of the API.
          type: string
        version:
          description: Version of the API.
          type: string
        description:
          description: Description of the API.
          type: string
        draft:
          type: boolean
          description: Whether or not the API is in draft.
        public:
          type: boolean
          description: Whether or not the API is public.
        running:
          type: boolean
          description: Whether or not the API is running.
        entrypoints:
          description: List of all the avaible endpoints to call the API.
          type: array
          items:
            type: string
        labels:
          description: List of labels linked to this API.
          type: array
          items:
            type: string
        owner:
          $ref: '#/components/schemas/User'
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        categories:
          description: List of categories this API belongs to.
          type: array
          items:
            type: string
        rating_summary:
          $ref: '#/components/schemas/RatingSummary'
        _links:
          $ref: '#/components/schemas/ApiLinks'

        ###################
        # may be included #
        ###################
        pages:
          description: |
            Only returned with */apis/{apiId}*.
            Need *include* query param to contain 'pages'.

            The documentation pages of this API. Same as */apis/{apiId}/pages*.
          type: array
          items:
            $ref: '#/components/schemas/Page'
        plans:
          description: |
            Only returned with */apis/{apiId}*.
            Need *include* query param to contain 'plans'.

            The plans of this API. Same as */apis/{apiId}/plans*.
          type: array
          items:
            $ref: '#/components/schemas/Plan'
    ApiInformation:
      properties:
        name:
          type: string
          description: Name of api information
        value:
          type: string
          description: Value of api information
...
    Page:
      required:
        - id
        - name
        - type
        - order
      properties:
        id:
          description: Unique identifier of a page.
          type: string
        name:
          description: Name of the page.
          type: string
        type:
          description: Type of documentation.
          type: string
          enum:
            - SWAGGER
            - MARKDOWN
            - FOLDER
            - ROOT
            - LINK
        order:
          description: Order of the documentation page in its folder.
          type: integer
        parent:
          description: Parent page. MAY be null.
          type: string
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        configuration:
          $ref: '#/components/schemas/PageConfiguration'
        metadata:
          description: Array of metadata about the page. This array is filled when the page has been fetched from a distant source (GitHub, GitLab, etc...).
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
        _links:
          $ref: '#/components/schemas/PageLinks'

        ###################
        # may be included #
        ###################
        content:
          description: |
            Only returned with */apis/{apiId}/pages/{pageId}* and */pages/{pageId}*.
            Need *include* query param to contain 'content'.

            The content of the page.
          type: string
    Plan:
      required:
        - id
        - name
        - security
        - type
        - description
        - validation
        - order
        - comment_required
      properties:
        id:
          description: Unique identifier of a plan.
          type: string
        name:
          description: Name of the plan.
          type: string
        security:
          description: Security used with this plan.
          type: string
          enum:
            - API_KEY
            - KEY_LESS
            - JWT
            - OAUTH2
        description:
          description: Description of the plan.
          type: string
        characteristics:
          description: List of additionnal terms to describe the plan.
          type: array
          items:
            type: string
        validation:
          description: Type of validation for subscription requests.
          type: string
          enum:
            - AUTO
            - MANUAL
        order:
          description: Priority order
          type: integer
        comment_required:
          description: True if a comment is required when a subscription is created.
          type: boolean
        comment_question:
          description: Content of the message sent to a user creating a subscription.
          type: string
    Rating:
      required:
        - id
        - title
        - value
        - comment
      properties:
        id:
          description: Unique identifier of a rating.
          type: string
        title:
          description: Title of a rating.
          type: string
        value:
          description: Value of the rating. Integer between 1 and 5.
          type: integer
        author:
          $ref: '#/components/schemas/User'
        date:
          description: Date and time of the rating.
          type: string
          format: date-time
        comment:
          description: Comment left by the author to explain the rating.
          type: string
        answers:
          description: Array of answers.
          type: array
          items:
            $ref: '#/components/schemas/RatingAnswer'
...
    User:
      properties:
        id:
          description: Unique identifier of a user.
          type: string
        reference:
          description: Unique reference if user comes from external source. Use for search only.
          type: string
        first_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
        config:
          $ref: '#/components/schemas/UserConfig'
        _links:
          $ref: '#/components/schemas/UserLinks'
    UserConfig:
      properties:
        management_url:
          description: The URL of the Gravitee management UI
          type: string
...
    CategorizedLinks:
      properties:
        root:
          type: boolean
          description: true if the links of this category are in system folder and not in a subfolder.
        category:
          type: string
          description: name of the group of links
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    RatingInput:
      properties:
        title:
          type: string
        value:
          type: integer
          minimum: 0
          maximum: 5
        comment:
          type: string
...
    Metadata:
      description: Information about the page, if this page is from an external source.
      properties:
        name:
          type: string
        value:
          type: string
        order:
          type: string
    RatingSummary:
      properties:
        average:
          type: number
          format: double
        count:
          type: number
        one_star:
          type: number
          format: double
        two_star:
          type: number
          format: double
        three_star:
          type: number
          format: double
        four_star:
          type: number
          format: double
        five_star:
          type: number
          format: double
...
    RatingAnswer:
      required:
        - id
        - author
        - date
        - comment
      properties:
        id:
          type: string
        author:
          $ref: '#/components/schemas/User'
        date:
          type: string
          format: date-time
        comment:
          type: string
...
    MetadataMap:
      description: Map of Map of Object
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: object
...
    Links:
      properties:
        self:
          type: string
        first:
          type: string
        last:
          type: string
        prev:
          type: string
        next:
          type: string
    ApiLinks:
      properties:
        self:
          type: string
        links:
          type: string
        metrics:
          type: string
        pages:
          type: string
        picture:
          type: string
        background:
          type: string
        plans:
          type: string
        ratings:
          type: string
...
    PageLinks:
      properties:
        self:
          type: string
        content:
          type: string
        parent:
          type: string
...
    UserLinks:
      properties:
        self:
          type: string
        avatar:
          type: string
        notifications:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    APINotFoundError:
      description: API not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PagesSuccess:
      description: List of pages
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PagesResponse'
    PictureSuccess:
      description: A picture
      content:
        'image/*':
          schema:
            type: string
            format: binary
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM